# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Homeschool Hive API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /links:
    get:
      operationId: listLinks
      tags: [links]
      summary: List educational links with optional search and filter
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [newest, oldest, most_reactions, title]
      responses:
        "200":
          description: List of links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LinkWithDetails"

    post:
      operationId: createLink
      tags: [links]
      summary: Create a new educational link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkRequest"
      responses:
        "201":
          description: Created link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkWithDetails"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /links/{id}:
    get:
      operationId: getLink
      tags: [links]
      summary: Get a single link by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Link details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkWithDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    put:
      operationId: updateLink
      tags: [links]
      summary: Update an existing link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLinkRequest"
      responses:
        "200":
          description: Updated link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkWithDetails"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    delete:
      operationId: deleteLink
      tags: [links]
      summary: Delete a link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /links/{id}/reactions:
    post:
      operationId: addReaction
      tags: [reactions]
      summary: Add or toggle an emoji reaction on a link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddReactionRequest"
      responses:
        "200":
          description: Updated reactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionsResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        lastNameInitial:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
      required:
        - id
...
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        color:
          type: string
      required:
        - id
        - name
        - slug
        - color
    Reaction:
      type: object
      properties:
        emoji:
          type: string
        count:
          type: integer
        userReacted:
          type: boolean
      required:
        - emoji
        - count
        - userReacted
    ReactionsResponse:
      type: object
      properties:
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/Reaction"
      required:
        - reactions
    Link:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        title:
          type: string
        summary:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userId:
          type: string
      required:
        - id
        - url
        - title
        - createdAt
        - updatedAt
        - userId
    LinkWithDetails:
      allOf:
        - $ref: "#/components/schemas/Link"
        - type: object
          properties:
            tags:
              type: array
              items:
                $ref: "#/components/schemas/Tag"
            reactions:
              type: array
              items:
                $ref: "#/components/schemas/Reaction"
            user:
              $ref: "#/components/schemas/AuthUser"
            totalReactions:
              type: integer
          required:
            - tags
            - reactions
            - user
            - totalReactions
    CreateLinkRequest:
      type: object
      properties:
        url:
          type: string
        title:
          type: string
        summary:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        tagIds:
          type: array
          items:
            type: integer
      required:
        - url
        - title
    UpdateLinkRequest:
      type: object
      properties:
        url:
          type: string
        title:
          type: string
        summary:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        tagIds:
          type: array
          items:
            type: integer
    AddReactionRequest:
      type: object
      properties:
        emoji:
          type: string
      required:
        - emoji