# lib/api-zod/src/generated/types/link.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Homeschool Hive API
 * OpenAPI spec version: 0.1.0
 */

export type LinkPriceRange = "free" | "under_10" | "under_50" | "under_100" | "under_500" | "over_500";

export interface Link {
  id: number;
  url: string;
  title: string;
  summary?: string | null;
  comment?: string | null;
  thumbnailUrl?: string | null;
  createdAt: Date;
  updatedAt: Date;
  userId: string;
  status?: "pending" | "approved" | "rejected" | null;
  priceRange?: LinkPriceRange | null;
}